Move semantic analyzer to core/semantic/ directory
- Copy xvr_semantic.h and xvr_semantic.c to src/core/semantic/
- Update includes to use relative paths (../ast/, ../types/)
- Create redirect files in src/ that include from core/
- Update CMakeLists.txt to compile from core/semantic/
